apps/adk/pyproject.toml
[project]
name = "adk"
version = "0.1.0"
description = "google ADK adventures"
readme = "README.md"
authors = [
    { name = "Sumanth Chinthagunta", email = "xmlking@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "google-adk>=0.1.0",
    "litellm>=1.66.0",
]

[tool.pyright]
venvPath = "../.."
venv = ".venv"
strict = ["**/*.py"]
pythonVersion = "3.12"

[tool.ruff]
extend = "../../pyproject.toml"
exclude = ["build", "dist"]
include = [
    "src/**"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["google_search","hello"]

